% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_row.R
\name{create_row}
\alias{create_row}
\title{helper function to create a new row for the task queue}
\usage{
create_row(
  id = 1L,
  state = "waiting",
  fun = function() NULL,
  args = NULL,
  result = NULL
)
}
\arguments{
\item{id}{integer with the task identifying index}

\item{state}{a string indicating the status of the job. Can be one of
"waiting", "idle", "running" or "done"}

\item{fun}{a function}

\item{args}{a list of arguments for the function}

\item{result}{a list wiht the results of a function. This is used for
reloading tasks from the DB.}
}
\value{
a tibble with one row
}
\description{
helper function to create a new row for the task queue
}
\examples{

caRdoon:::create_row()
# A tibble: 1 × 7
#    id idle  state   fun    args       worker result
#  <int> <lgl> <chr>   <list> <list>     <list> <list>
#1     1 FALSE waiting <fn>   <list [1]> <NULL> <NULL>

}
