% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_task_object.R
\name{create_task_object}
\alias{create_task_object}
\title{Create R6 task list object}
\usage{
create_task_object(num_worker = 1L)
}
\arguments{
\item{num_worker}{integer, the number of worker processes}
}
\value{
an R6 object to track and queue tasks
}
\description{
The process and details for this R6 class can be found
  (here)[https://www.tidyverse.org/blog/2019/09/callr-task-q/].
}
