% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{db_update_row}
\alias{db_update_row}
\title{Send UPDATE SQL statement}
\usage{
db_update_row(con, tablename, key, set)
}
\arguments{
\item{con}{connection to database}

\item{tablename}{name of the table in DB}

\item{key}{a named list with key parameters, where the name is the column and
the value the value. Used in WHERE SQL-statement.}

\item{set}{a named list with update parameters, where the name is the column
and the value the value. Used in SET SQL-statement.}
}
\value{
a Boolean that indicates if the SQL statement was successful.
}
\description{
Send UPDATE SQL statement
}
