% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_sql_value}
\alias{get_sql_value}
\title{Extracts values from list for SQL statement}
\usage{
get_sql_value(x)
}
\arguments{
\item{x}{named list, where the name is the column and the value the value}
}
\value{
a string with name-value pairs separated by commas.
}
\description{
Extracts values from list for SQL statement
}
