% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_value}
\alias{get_value}
\title{Add quotations on string values}
\usage{
get_value(x)
}
\arguments{
\item{x}{an object of length one, either a numeric or a character}
}
\value{
a string
}
\description{
Add quotations on string values
}
