% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cardoon.R
\name{run_cardoon}
\alias{run_cardoon}
\title{Start up caRdoon API}
\usage{
run_cardoon(port = 9662)
}
\arguments{
\item{port}{integer with the port, the API should run on}
}
\value{
a message, that the API has closed
}
\description{
Running this function will create the caRdoon API.
The API acts as a task queue with multiple workers in the background.
}
