% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cardoon.R
\name{run_cardoon}
\alias{run_cardoon}
\title{Start up caRdoon API}
\usage{
run_cardoon(port = 9662, num_worker = 1, docs = FALSE)
}
\arguments{
\item{port}{integer with the port, the API should run on}

\item{num_worker}{integer with number of worker processes}

\item{docs}{a boolean indicating if the docs should be started}
}
\value{
a message, that the API has closed
}
\description{
Running this function will create the caRdoon API.
The API acts as a task queue with multiple workers in the background.
}
